refactoring varbase in the same way as periodbase...
