+++ astrobase/varbase/lcfit.py
@@ -1,0 +1,396 @@
+#####################################################
+## FOURIER FITTING TO PHASED MAGNITUDE TIME SERIES ##
+#####################################################
+
+def _fourier_func(fourierparams, phase, mags):
+    '''
+    This returns a summed Fourier series generated using fourierparams.
+
+    fourierparams is a sequence like so:
+
+    [ampl_1, ampl_2, ampl_3, ..., ampl_X, pha_1, pha_2, pha_3, ..., pha_X]
+
+    where X is the Fourier order.
+
+    mags and phase MUST have no nans.
+
+    '''
+
+    # figure out the order from the length of the Fourier param list
+    order = len(fourierparams)/2
+
+    # get the amplitude and phase coefficients
+    f_amp = fourierparams[:order]
+    f_pha = fourierparams[order:]
+
+    # calculate all the individual terms of the series
+    f_orders = [f_amp[x]*npcos(2.0*MPI*x*phase + f_pha[x])
+                for x in range(order)]
+
+    # this is the zeroth order coefficient - a constant equal to median mag
+    total_f = npmedian(mags)
+
+    # sum the series
+    for fo in f_orders:
+        total_f += fo
+
+    return total_f
+
+
+
+def _fourier_chisq(fourierparams,
+                   phase,
+                   mags,
+                   errs):
+    '''
+    This is the chisq objective function to be minimized by scipy.minimize.
+
+    The parameters are the same as _fourier_func above.
+
+    '''
+
+    f = _fourier_func(fourierparams, phase, mags)
+    chisq = npsum(((mags - f)*(mags - f))/(errs*errs))
+
+    return chisq
+
+
+
+def _fourier_residual(fourierparams,
+                      phase,
+                      mags):
+    '''
+    This is the residual objective function to be minimized by scipy.leastsq.
+
+    The parameters are the same as _fourier_func above.
+
+    '''
+
+    f = _fourier_func(fourierparams, phase, mags)
+    residual = mags - f
+
+    return residual
+
+
+
+def fourier_fit_magseries(times, mags, errs, period,
+                          initfourierparams=[0.6,0.2,0.2,0.2,0.2,0.2,0.2,0.2,
+                                             0.1,0.1,0.1,0.1,0.1,0.1,0.1,0.1],
+                          sigclip=3.0,
+                          plotfit=False,
+                          ignoreinitfail=True):
+    '''This fits a Fourier series to a magnitude time series.
+
+    This uses an 8th-order Fourier series by default. This is good for light
+    curves with many thousands of observations (HAT light curves have ~10k
+    observations). Lower the order accordingly if you have less observations in
+    your light curves to avoid over-fitting.
+
+    Returns the Fourier fit parameters, the minimum chisq and reduced
+    chisq. Makes a plot for the fit to the mag series if plotfit is a string
+    containing a filename to write the plot to.
+
+    This folds the time series using the given period and at the first
+    observation. Can optionally sigma-clip observations.
+
+    if ignoreinitfail is True, ignores the initial failure to find a set of
+    optimized Fourier parameters and proceeds to do a least-squares fit anyway.
+
+    '''
+
+    # get rid of nans first
+    find = npisfinite(times) & npisfinite(mags) & npisfinite(errs)
+    ftimes = times[find]
+    fmags = mags[find]
+    ferrs = errs[find]
+
+    # get the median and stdev = 1.483 x MAD
+    median_mag = npmedian(fmags)
+    stddev_mag = (npmedian(npabs(fmags - median_mag))) * 1.483
+
+    # sigclip next
+    if sigclip:
+
+        sigind = (npabs(fmags - median_mag)) < (sigclip * stddev_mag)
+
+        stimes = ftimes[sigind]
+        smags = fmags[sigind]
+        serrs = ferrs[sigind]
+
+        LOGINFO('sigclip = %s: before = %s observations, '
+                'after = %s observations' %
+                (sigclip, len(times), len(stimes)))
+
+    else:
+
+        stimes = ftimes
+        smags = fmags
+        serrs = ferrs
+
+    # phase the mag series using the given period and faintest mag time
+    # mintime = stimes[npwhere(smags == npmax(smags))]
+
+    # phase the mag series using the given period and epoch = min(stimes)
+    mintime = npmin(stimes)
+
+    # calculate the unsorted phase, then sort it
+    iphase = (stimes - mintime)/period - npfloor((stimes - mintime)/period)
+    phasesortind = npargsort(iphase)
+
+    # these are the final quantities to use for the Fourier fits
+    phase = iphase[phasesortind]
+    pmags = smags[phasesortind]
+    perrs = serrs[phasesortind]
+
+    # get the times sorted in phase order (useful to get the fit mag minimum
+    # with respect to phase -- the light curve minimum)
+    ptimes = stimes[phasesortind]
+
+    fourierorder = len(initfourierparams)/2
+
+    LOGINFO('fitting Fourier series of order %s to '
+            'mag series with %s observations, '
+            'using period %.6f, folded at %.6f' % (fourierorder,
+                                                   len(phase),
+                                                   period,
+                                                   mintime))
+
+    # initial minimize call to find global minimum in chi-sq
+    initialfit = spminimize(_fourier_chisq,
+                            initfourierparams,
+                            method='BFGS',
+                            args=(phase, pmags, perrs))
+
+    # make sure this initial fit succeeds before proceeding
+    if initialfit.success or ignoreinitfail:
+
+        LOGINFO('initial fit done, refining...')
+
+        leastsqparams = initialfit.x
+
+        leastsqfit = spleastsq(_fourier_residual,
+                               leastsqparams,
+                               args=(phase, pmags))
+
+        # if the fit succeeded, then we can return the final parameters
+        if leastsqfit[-1] in (1,2,3,4):
+
+            finalparams = leastsqfit[0]
+
+            # calculate the chisq and reduced chisq
+            fitmags = _fourier_func(finalparams, phase, pmags)
+
+            fitchisq = npsum(
+                ((fitmags - pmags)*(fitmags - pmags)) / (perrs*perrs)
+            )
+
+            fitredchisq = fitchisq/(len(pmags) - len(finalparams) - 1)
+
+            LOGINFO(
+                'final fit done. chisq = %.5f, reduced chisq = %.5f' %
+                (fitchisq,fitredchisq)
+            )
+
+            # figure out the time of light curve minimum (i.e. the fit epoch)
+            # this is when the fit mag is maximum (i.e. the faintest)
+            fitmagminind = npwhere(fitmags == npmax(fitmags))
+            magseriesepoch = ptimes[fitmagminind]
+
+            # assemble the returndict
+            returndict =  {'finalparams':finalparams,
+                           'initialfit':initialfit,
+                           'leastsqfit':leastsqfit,
+                           'fitchisq':fitchisq,
+                           'fitredchisq':fitredchisq,
+                           'fitplotfile':None,
+                           'phase':phase,
+                           'mags':pmags,
+                           'errs':perrs,
+                           'fitmags':fitmags,
+                           'fitepoch':magseriesepoch}
+
+            # make the fit plot if required
+            if plotfit and isinstance(plotfit, str):
+
+                plt.figure(figsize=(8,6))
+                plt.axvline(0.5,color='g',linestyle='--')
+                plt.errorbar(phase,pmags,fmt='bo',yerr=perrs,
+                             markersize=2.0,capsize=0)
+                plt.plot(phase,fitmags, 'r-',linewidth=2.0)
+                ymin, ymax = plt.ylim()
+                plt.ylim(ymax,ymin)
+                plt.gca().set_xticks(
+                    [0.0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1.0]
+                )
+                plt.xlabel('phase')
+                plt.ylabel('magnitude')
+                plt.title('period: %.6f, folded at %.6f, fit epoch: %.6f' %
+                          (period, mintime, magseriesepoch))
+                plt.savefig(plotfit)
+                plt.close()
+
+                returndict['fitplotfile'] = plotfit
+
+            return returndict
+
+
+    # if the fit didn't succeed, we can't proceed
+    else:
+
+        LOGERROR('initial Fourier fit did not succeed, '
+                 'reason: %s, returning scipy OptimizeResult'
+                 % initialfit.message)
+
+        return {'finalparams':None,
+                'initialfit':initialfit,
+                'leastsqfit':None,
+                'fitchisq':None,
+                'fitredchisq':None,
+                'fitplotfile':None,
+                'phase':phase,
+                'mags':pmags,
+                'errs':perrs,
+                'fitmags':None,
+                'fitepoch':None}
+
+
+#################################################################
+## SPLINE FITTING TO PHASED AND UNPHASED MAGNITUDE TIME SERIES ##
+#################################################################
+
+def spline_fit_magseries(times, mags, errs, period,
+                         knotfraction=0.01,
+                         maxknots=100,
+                         sigclip=30.0,
+                         plotfit=False,
+                         ignoreinitfail=False):
+
+    '''This fits a univariate cubic spline to the phased light curve.
+
+    This fit may be better than the Fourier fit for sharply variable objects,
+    like EBs, so can be used to distinguish them from other types of variables.
+
+    The knot fraction is the number of internal knots to use for the spline. A
+    value of 0.01 (or 1%) of the total number of non-nan observations appears to
+    work quite well, without over-fitting.
+
+    Returns the chisq of the fit, as well as the reduced chisq. FIXME: check
+    this equation below to see if it's right.
+
+    reduced_chisq = fit_chisq/(len(pmags) - len(knots) - 1)
+
+    '''
+
+    if errs is None:
+        errs = npfull_like(mags, 0.005)
+
+    # get rid of nans first
+    find = npisfinite(times) & npisfinite(mags) & npisfinite(errs)
+    ftimes = times[find]
+    fmags = mags[find]
+    ferrs = errs[find]
+
+    # get the median and stdev = 1.483 x MAD
+    median_mag = npmedian(fmags)
+    stddev_mag = (npmedian(npabs(fmags - median_mag))) * 1.483
+
+    # sigclip next
+    if sigclip:
+
+        sigind = (npabs(fmags - median_mag)) < (sigclip * stddev_mag)
+
+        stimes = ftimes[sigind]
+        smags = fmags[sigind]
+        serrs = ferrs[sigind]
+
+        LOGINFO('sigclip = %s: before = %s observations, '
+                'after = %s observations' %
+                (sigclip, len(times), len(stimes)))
+
+    else:
+
+        stimes = ftimes
+        smags = fmags
+        serrs = ferrs
+
+    # phase the mag series using the given period and faintest mag time
+    # mintime = stimes[npwhere(smags == npmax(smags))]
+
+    # phase the mag series using the given period and epoch = min(stimes)
+    mintime = npmin(stimes)
+
+    # calculate the unsorted phase, then sort it
+    iphase = (stimes - mintime)/period - npfloor((stimes - mintime)/period)
+    phasesortind = npargsort(iphase)
+
+    # these are the final quantities to use for the Fourier fits
+    phase = iphase[phasesortind]
+    pmags = smags[phasesortind]
+    perrs = serrs[phasesortind]
+
+    # get the times sorted in phase order (useful to get the fit mag minimum
+    # with respect to phase -- the light curve minimum)
+    ptimes = stimes[phasesortind]
+
+    # now figure out the number of knots up to max knots (=100)
+    nobs = len(phase)
+    nknots = int(npfloor(knotfraction*nobs))
+    nknots = maxknots if nknots > maxknots else nknots
+    splineknots = nplinspace(phase[0]+0.01,phase[-1]-0.01,num=nknots)
+
+    # generate and fit the spline
+    spl = LSQUnivariateSpline(phase,pmags,t=splineknots,w=1.0/perrs)
+
+    # calculate the spline fit to the actual phases, the chisq and red-chisq
+    fitmags = spl(phase)
+
+    fitchisq = npsum(
+        ((fitmags - pmags)*(fitmags - pmags)) / (perrs*perrs)
+    )
+
+    fitredchisq = fitchisq/(len(pmags) - nknots - 1)
+
+    LOGINFO(
+        'spline fit done. nknots = %s,  chisq = %.5f, reduced chisq = %.5f' %
+        (nknots, fitchisq, fitredchisq)
+    )
+
+    # figure out the time of light curve minimum (i.e. the fit epoch)
+    # this is when the fit mag is maximum (i.e. the faintest)
+    fitmagminind = npwhere(fitmags == npmax(fitmags))
+    magseriesepoch = ptimes[fitmagminind]
+
+    # assemble the returndict
+    returndict =  {'nknots':nknots,
+                   'fitchisq':fitchisq,
+                   'fitredchisq':fitredchisq,
+                   'fitplotfile':None,
+                   'phase':phase,
+                   'mags':pmags,
+                   'errs':perrs,
+                   'fitmags':fitmags,
+                   'fitepoch':magseriesepoch}
+
+    # make the fit plot if required
+    if plotfit and isinstance(plotfit, str):
+
+        plt.figure(figsize=(8,6))
+        plt.axvline(0.5,color='g',linestyle='--')
+        plt.errorbar(phase,pmags,fmt='bo',yerr=perrs,
+                     markersize=2.0,capsize=0)
+        plt.plot(phase,fitmags, 'r-',linewidth=2.0)
+        ymin, ymax = plt.ylim()
+        plt.ylim(ymax,ymin)
+        plt.gca().set_xticks(
+            [0.0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1.0]
+        )
+        plt.xlabel('phase')
+        plt.ylabel('magnitude')
+        plt.title('period: %.6f, folded at %.6f, fit epoch: %.6f' %
+                  (period, mintime, magseriesepoch))
+        plt.savefig(plotfit)
+        plt.close()
+
+        returndict['fitplotfile'] = plotfit
+
+    return returndict
